+++ src/HomeAutio.Mqtt.GoogleHome/Models/Schema/CommandSchema.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using HomeAutio.Mqtt.GoogleHome.Extensions;
 using Newtonsoft.Json;
 using NJsonSchema;
