+++ src/HomeAutio.Mqtt.GoogleHome/Models/Schema/StateSchema.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using HomeAutio.Mqtt.GoogleHome.Extensions;
 using HomeAutio.Mqtt.GoogleHome.Models.State;
 using Newtonsoft.Json;
 using NJsonSchema;
