+++ src/HomeAutio.Mqtt.GoogleHome/Models/State/Device.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using HomeAutio.Mqtt.GoogleHome.Extensions;
 using Newtonsoft.Json;
 
 namespace HomeAutio.Mqtt.GoogleHome.Models.State
