+++ src/HomeAutio.Mqtt.GoogleHome/Views/GoogleDevice/Edit.cshtml
@@ -1,4 +1,4 @@
-@using HomeAutio.Mqtt.GoogleHome
+@using HomeAutio.Mqtt.GoogleHome.Extensions
 @model HomeAutio.Mqtt.GoogleHome.ViewModels.DeviceViewModel
 
 @section Scripts {
