+++ src/HomeAutio.Mqtt.GoogleHome/Views/GoogleTrait/Edit.cshtml
@@ -1,4 +1,4 @@
-@using HomeAutio.Mqtt.GoogleHome
+@using HomeAutio.Mqtt.GoogleHome.Extensions
 @model HomeAutio.Mqtt.GoogleHome.ViewModels.TraitViewModel
 
 @section Scripts {
